+++ alsek/storage/backends/postgres/standard/backend.py
@@ -17,7 +17,7 @@ from sqlalchemy.orm import Session
 
 from alsek.defaults import DEFAULT_NAMESPACE
 from alsek.storage.backends import Backend, LazyClient
-from alsek.storage.backends.postgres._utils import (
+from alsek.storage.backends.postgres.utils.general import (
     validate_value_within_postgres_notification_size_limit,
 )
 from alsek.storage.backends.postgres.standard._pubsub_listener import PostgresPubSubListener
