+++ alsek/storage/backends/postgres/utils/pubsub.py
@@ -1,0 +1,34 @@
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import Any, Iterable, AsyncIterable
+
+from sqlalchemy import URL
+
+from alsek.storage.serialization import Serializer
+
+
+class BasePostgresPubSubListen(ABC):
+    def __init__(
+        self,
+        channel: str,
+        url: URL,
+        serializer: Serializer,
+        sleep_time: float = 0.01,
+    ) -> None:
+        self.channel = channel
+        self.url = url
+        self.serializer = serializer
+        self.sleep_time = sleep_time
+
+    @staticmethod
+    def _parse_notification_data(
+        payload: str,
+        serializer: Serializer,
+    ) -> dict[str, Any]:
+        data = serializer.reverse(payload)
+        return {"type": "message", "data": data}
+
+    @abstractmethod
+    def listen(self) -> Iterable[Any] | AsyncIterable[Any]:
+        raise NotImplementedError
